fix: rename usage to instrumentation
